chore(scripts): migrate gh actions scripts - I4
